issue create

+++ src/main/java/ai/group2/project_management_system/model/entity/Department.java
@@ -18,7 +18,8 @@ public class Department {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+
+    private Long id;
     private String name;
     private String email;
     private String phone;
