+++ src/main/java/ai/group2/project_management_system/service/Impl/DepartmentServiceImpl.java
@@ -25,7 +25,7 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public Department getDepartmentById(Integer id) {
+    public Department getDepartmentById(Long id) {
         return departmentRepository.getReferenceById(id);
     }
